requirements.txt
anthropic
flask
python-dotenv
python-docx
python-pptx
openpyxl
